Gestion en comite y concejo de reconocimiento de creditos por actividades practica docente

+++ src/main/java/com/maestria/gestionSolicitudes/dto/rest/request/DatosActividadDocenteRequest.java
@@ -1,5 +1,6 @@
 package com.maestria.gestionSolicitudes.dto.rest.request;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
@@ -8,7 +9,7 @@ import lombok.Data;
 public class DatosActividadDocenteRequest {
     private String codigoSubtipo;
     private Integer intensidadHoraria;
-    private Integer horasReconocer;
+    private BigDecimal horasReconocer;
     private List<String> documentosAdjuntos;
     private List<String> enlacesAdjuntos;
 }
