+++ src/main/java/com/maestria/gestionSolicitudes/dto/rest/response/DatosActividadDocenteResponse.java
@@ -1,5 +1,6 @@
 package com.maestria.gestionSolicitudes.dto.rest.response;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
@@ -7,7 +8,7 @@ import lombok.Data;
 @Data
 public class DatosActividadDocenteResponse {
     private String nombreActividad;
-    private Integer horasReconocer;
+    private BigDecimal horasReconocer;
     private List<String> documentos;
     private List<String> enlaces;
 }
